Ensures that the root cryptid file does not load the settings
