+++ src/bin/cryptid-logout.js
@@ -1,8 +1,3 @@
 import {SETTINGS} from './cli';
 
-SETTINGS.token = '';
-SETTINGS.email = '';
-SETTINGS.loggedIn = false;
-delete SETTINGS.checkLogin;
-
-process.exit();
+SETTINGS.clear();
