+++ src/bin/cryptid-properties.js
@@ -1,4 +1,4 @@
-import {loadHelp} from './cli';
+import {loadHelp} from './help';
 import program from 'commander';
 
 program
